fixed graph jitter on filter text input
