--- todos.js
@@ -1,19 +1,0 @@
-styling for job calls filter
-	some kind of headline for the job calls viewer
-
-
-
-pagination for results in job calls viewer
-
-uneccessary rendering?
-	changing text in filter causes re-render of chart components
-
-update database records
-
-deployment
-
-// DONE //========================================================//
-
-// if results are for a company, display the company along with the date range
-
-//styling for Displaying results component
